changed bonusScore to Double in all cases for recommendation system and improved junit tests

+++ src/main/java/eu/fbk/das/rs/sortfilter/DifficultyPrizeComparator.java
@@ -10,10 +10,10 @@ public class DifficultyPrizeComparator implements Comparator<ChallengeDataDTO> {
 	public int compare(ChallengeDataDTO o1, ChallengeDataDTO o2) {
 		int difficulty1 = (int) o1.getData().get("difficulty");
 		int difficulty2 = (int) o2.getData().get("difficulty");
-		long bonusScore1 = (long) o1.getData().get("bonusScore");
-		long bonusScore2 = (long) o2.getData().get("bonusScore");
-		double wi1 = (double) o1.getData().get("wi");
-		double wi2 = (double) o2.getData().get("wi");
+		Double bonusScore1 = (Double) o1.getData().get("bonusScore");
+		Double bonusScore2 = (Double) o2.getData().get("bonusScore");
+		Double wi1 = (Double) o1.getData().get("wi");
+		Double wi2 = (Double) o2.getData().get("wi");
 
 		if (difficulty1 < difficulty2) {
 			return -1;
